### 2016-04-12 #36 @masalmon
opencage -- a R package for the opencage API (forward and reverse geocoding)
* 1. What does this package do? (explain in 50 words or less)
This package is an interface to the opencage API (free account: 2500 calls/day). It allows forward geocoding (from placename to lat/lon) and reverse geocoding (vice versa). The API has many data sources.
* 2. Paste the full DESCRIPTION file inside a code block below.
```r
Package: opencage
Title: Interface To The OpenCage API
Version: 0.1.0
Authors@R: person("Maëlle", "Salmon", email = "maelle.salmon@yahoo.se", role = c("aut", "cre"))
Description: This packages accesses the OpenCage API which provides Forward
    geocoding (from placename to longitude and latitude) and Reverse geocoding (from
    longitude and latitude to placename)
    httr,
    memoise
Suggests: testthat, lintr
```
* 3. URL for the package (the development repository, not a stylized html page)
https://github.com/masalmon/opencage
* 4. What data source(s) does it work with (if applicable)?
opencage website states "open geospatial data including OpenStreetMap, Yahoo! GeoPlanet, Natural Earth Data, Thematic Mapping, Ordnance Survey OpenSpace, Statistics New Zealand, Zillow, MaxMind, GeoNames, the US Census Bureau and Flickr's shapefiles plus a whole lot more besides."
* 5. Who is the target audience?
People that need to geocode.
* 6. Are there other R packages that accomplish the same thing? If so, what is different about yours?
I know no function for reverse geocoding but for forward geocoding there's the geocode function in ggmap. "Geocodes a location (find latitude and longitude) using either (1) the Data Science Toolkit (http://www.datasciencetoolkit.org/about) or (2) Google Maps." However it has either Google Maps or Data Science Toolkit. opencage actually builds on other geocoders including Data Science Toolkit ("There's Nominatim, Data Science Toolkit and the Two Fishes geocoders." )

* 7. Check the box next to each policy below, confirming that you agree. These are mandatory.
    * [ x] This package does not violate the Terms of Service of any service it interacts with.
    * [ x] The repository has continuous integration with Travis CI and/or another service
    * [ ] The package contains a vignette
    * [ x] The package contains a reasonably complete README with `devtools` install instructions
    * [ x] The package contains unit tests
    * [ x] The package only exports functions to the NAMESPACE that are intended for end users
* 8. Do you agree to follow the [rOpenSci packaging guidelines](https://github.com/ropensci/packaging_guide)? These aren't mandatory, but we strongly suggest you follow them. If you disagree with anything, please explain.
    * [ ] Are there any package dependencies not on CRAN?
    * [ x] Do you intend for this package to go on CRAN?
    * [ x] Does the package have a CRAN accepted license?
    * [ ] Did `devtools::check()` produce any errors or warnings? If so paste them below.
* 9. Please add explanations below for any exceptions to the above:

The package does not have a vignette yet but it only has two functions with nearly the same arguments. However I agree that it might be good to show use cases, but since I don't geocode a lot in R, I need a bit of feedback about what applications are for users that really geocode.

* 10. If this is a resubmission following rejection, please explain the change in circumstances.
