### 2016-04-15 7f69ae6
adds rate info
