+++ R/utils.R
@@ -32,7 +32,12 @@ opencage_parse <- function(req) {
   list(results = results,
        total_results = no_results,
        time_stamp = lubridate::dmy_hms(temp$timestamp$created_http,
-                                       tz = "GMT"))
+                                       tz = "GMT"),
+       rate_info = dplyr::tbl_df(data.frame(
+         limit = temp$rate$limit,
+         remaining = temp$rate$remaining,
+         rest = as.POSIXct(temp$rate$reset, origin="1970-01-01")
+       )))
 }
 
 # base URL for all queries
