+++ man/opencage_reverse.Rd
@@ -35,6 +35,7 @@ A list with
 \item results as a data.frame (`dplyr` `tbl_df`) called results with one line per results,
 \item the number of results as an integer
 \item the timestamp as a POSIXct object
+\item rate_info data.frame (`dplyr` `tbl_df`) with the maximal number of API calls per day for the used key, the number of remaining calls for the day and the time at which the number of remaining calls with be reset.
 }
 }
 \description{
