### 2016-04-15 586e8aa
improves Travis config file
