### 2016-04-15 1eea651
updates readme, improves doc and changes + to , for lat/lon which is better
