+++ R/reverse_geocoding.R
@@ -28,7 +28,7 @@
 
   # res
   temp <- opencage_get(query_par = list(q = paste0(latitude,
-                                                   "+", longitude),
+                                                   ",", longitude),
                                         key = key,
                                         bounds = bounds,
                                         countrycode = countrycode,
@@ -55,7 +55,7 @@
 #' @param longitude longitude
 #' @param key your OpenCage key
 #' @param bounds Provides the geocoder with a hint to the region that the query resides in. This value will restrict the possible results to the supplied region. The bounds parameter should be specified as 4 coordinate points forming the south-west and north-east corners of a boundsing box. For example bounds=-0.563160,51.280430,0.278970,51.683979 (min long, min lat, max long, max lat).
-#' @param countrycode Restricts the results to the given country. The country code is a two letter code as defined by the ISO 3166-1 Alpha 2 standard. E.g. 'gb' for the United Kingdom, 'fr' for France, 'us' for United States.
+#' @param countrycode Restricts the results to the given country. The country code is a two letter code as defined by the ISO 3166-1 Alpha 2 standard. E.g. 'GB' for the United Kingdom, 'FR' for France, 'US' for United States.
 #' @param language an IETF format language code (such as es for Spanish or pt-BR for Brazilian Portuguese). If no language is explicitly specified, we will look for an HTTP Accept-Language header like those sent by a brower and use the first language specified and if none are specified en (English) will be assumed
 #' @param limit How many results should be returned (1-100). Default is 10.
 #' @param min_confidence an integer from 1-10. Only results with at least this confidence will be returned.
