+++ README.md
@@ -20,7 +20,7 @@ Geocoding
 
 The [OpenCage](https://geocoder.opencagedata.com/) API supports forward and reverse geocoding. Sources of OpenCage are open geospatial data including OpenStreetMap, Yahoo! GeoPlanet, Natural Earth Data, Thematic Mapping, Ordnance Survey OpenSpace, Statistics New Zealand, Zillow, MaxMind, GeoNames, the US Census Bureau and Flickr's shapefiles plus a whole lot more besides. See [this page](https://geocoder.opencagedata.com/credits) for the full list of credits.
 
-Below are two simple examples. For more information about the query parameters, see the package documentation, the [API doc](https://geocoder.opencagedata.com/api) and [OpenCage FAQ](https://geocoder.opencagedata.com/faq).
+Below are two simple examples.
 
 Forward geocoding
 -----------------
@@ -29,41 +29,29 @@ Forward geocoding is from placename to latitude and longitude tuplet(s).
 
 ``` r
 library("opencage")
-opencage_forward(placename = "Sarzeau", key = Sys.getenv("OPENCAGE_KEY"))
+output <- opencage_forward(placename = "Sarzeau", key = Sys.getenv("OPENCAGE_KEY"))
+print(output$time_stamp)
 ```
 
-    ## $results
-    ## Source: local data frame [2 x 42]
-    ## 
-    ##    annotations.DMS.lat annotations.DMS.lng annotations.MGRS
-    ##                  (chr)               (chr)            (chr)
-    ## 1 47° 31' 43.56984'' N 2° 45' 51.11892'' W  30TWT1774963954
-    ## 2 47° 31' 40.80828'' N  2° 46' 7.68180'' W  30TWT1740363867
-    ## Variables not shown: annotations.Maidenhead (chr), annotations.Mercator.x
-    ##   (chr), annotations.Mercator.y (chr), annotations.OSM.edit_url (chr),
-    ##   annotations.OSM.url (chr), annotations.callingcode (fctr),
-    ##   annotations.geohash (chr), annotations.sun.rise.apparent (fctr),
-    ##   annotations.sun.rise.astronomical (fctr), annotations.sun.rise.civil
-    ##   (fctr), annotations.sun.rise.nautical (fctr),
-    ##   annotations.sun.set.apparent (fctr), annotations.sun.set.astronomical
-    ##   (fctr), annotations.sun.set.civil (fctr), annotations.sun.set.nautical
-    ##   (fctr), annotations.timezone.name (fctr),
-    ##   annotations.timezone.now_in_dst (fctr), annotations.timezone.offset_sec
-    ##   (fctr), annotations.timezone.offset_string (fctr),
-    ##   annotations.timezone.short_name (fctr), annotations.what3words.words
-    ##   (chr), bounds.northeast.lat (chr), bounds.northeast.lng (chr),
-    ##   bounds.southwest.lat (chr), bounds.southwest.lng (chr), components.city
-    ##   (fctr), components.country (fctr), components.country_code (fctr),
-    ##   components.county (fctr), components.postcode (fctr), components.state
-    ##   (fctr), confidence (chr), formatted (chr), geometry.lat (chr),
-    ##   geometry.lng (chr), components.post_office (fctr), components.road
-    ##   (fctr), components.suburb (fctr), components.village (fctr)
-    ## 
-    ## $total_results
-    ## [1] 2
-    ## 
-    ## $time_stamp
-    ## [1] "2016-04-10 07:03:59 UTC"
+    ## [1] "2016-04-15 21:07:55 GMT"
+
+``` r
+library("dplyr")
+output$rate_info %>% knitr::kable()
+```
+
[…]
