+++ vignettes/opencage.Rmd
@@ -1,0 +1,85 @@
+---
+title: "R package for the OpenCage API -- both forward and reverse geocoding"
+author: "Maëlle Salmon"
+date: "`r Sys.Date()`"
+output: rmarkdown::html_vignette
+vignette: >
+  %\VignetteIndexEntry{Vignette Title}
+  %\VignetteEngine{knitr::rmarkdown}
+  %\VignetteEncoding{UTF-8}
+---
+
+This package is an interface to the OpenCage API that allows forward and reverse geocoding. To use the package, you will also need an API key. For this register at https://geocoder.opencagedata.com/pricing. The free API key provides up to 2,500 calls a day. It is recommended you save your API key as an environment variable. See https://stat545-ubc.github.io/bit003_api-key-env-var.html
+
+# Geocoding
+
+The [OpenCage](https://geocoder.opencagedata.com/) API supports forward and reverse geocoding. Sources of OpenCage are open geospatial data including OpenStreetMap, Yahoo! GeoPlanet, Natural Earth Data, Thematic Mapping, Ordnance Survey OpenSpace, Statistics New Zealand, Zillow, MaxMind, GeoNames, the US Census Bureau and Flickr's shapefiles plus a whole lot more besides. See [this page](https://geocoder.opencagedata.com/credits) for the full list of credits.
+
+Below are two simple examples.
+
+## Forward geocoding
+
+Forward geocoding is from placename to latitude and longitude tuplet(s).
+
+```{r, warnings = FALSE, message = FALSE}
+library("opencage")
+output <- opencage_forward(placename = "Sarzeau", key = Sys.getenv("OPENCAGE_KEY"))
+print(output$time_stamp)
+library("dplyr")
+output$rate_info %>% knitr::kable()
+output$results %>% knitr::kable()
+```
+
+## Reverse geocoding
+
+Reverse geocoding is from latitude and longitude to placename(s).
+
+```{r}
+output2 <- opencage_reverse(latitude = 51.5034070, 
+                            longitude = -0.1275920, 
+                 key = Sys.getenv("OPENCAGE_KEY"))
+print(output2$time_stamp)
+output2$rate_info %>% knitr::kable()
+output2$results %>% knitr::kable()
+```
+
+## Output
+
+For both `opencage_forward` and `opencage_reverse` functions, the package returns a list with a time stamp for the query, the total number of results, a data.frame (`dplyr tbl_df`) with information about the remaining calls to the API unless you have an unlimited account, and a data.frame (`dplyr tbl_df`) with the results corresponding to your query. You can find longitude and latitude for each results as `geometry.lat` and `geometry.lng`. Other information includes country and country information, time of sunset and sunrise, geohash (a geocoding system identifying a point with a single string, as explained in many more details [here](https://www.elastic.co/guide/en/elasticsearch/guide/current/geohashes.html) and [here](https://en.wikipedia.org/wiki/Geohash) -- for pure conversion between longitude/latitude and geohashes, see [this package](https://github.com/Ironholds/geohash)).  Depending on the data available in the API for the results one gets different columns: there can be a lot to explore!
+
+## Parameters
+
+Optional parameters of both `opencage_forward` and `opencage_reverse` can make the query more precise:
+
+* `bounds`: Provides the geocoder with a hint to the region that the query resides in. This value will restrict the possible results to the supplied region. The bounds parameter should be specified as 4 coordinate points forming the south-west and north-east corners of a boundsing box. For example bounds=-0.563160,51.280430,0.278970,51.683979 (min long, min lat, max long, max lat).
+
[…]
