### 2016-04-25 4e665e8
Automatically look up API key in OPENCAGE_KEY
