+++ man/opencage_key.Rd
@@ -1,0 +1,14 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/utils.R
+\name{opencage_key}
+\alias{opencage_key}
+\title{Retrieve Opencage API key}
+\usage{
+opencage_key(quiet = FALSE)
+}
+\description{
+An Opencage API Key
+Looks in env var \code{OPENCAGE_KEY}
+}
+\keyword{internal}
+
