### 2016-04-26 8c00d49
corrects transformation of lat and lng
