+++ README.md
@@ -30,28 +30,38 @@ Forward geocoding is from placename to latitude and longitude tuplet(s).
 ``` r
 library("opencage")
 output <- opencage_forward(placename = "Sarzeau", key = Sys.getenv("OPENCAGE_KEY"))
+```
+
+    ## Warning: All formats failed to parse. No formats found.
+
+``` r
 print(output$time_stamp)
 ```
 
-    ## [1] "2016-04-15 21:07:55 GMT"
+    ## [1] NA
 
 ``` r
 library("dplyr")
+```
+
+    ## Warning: package 'dplyr' was built under R version 3.2.5
+
+``` r
 output$rate_info %>% knitr::kable()
 ```
 
 |  limit|  remaining| rest                |
 |------:|----------:|:--------------------|
-|   2500|       2325| 2016-04-16 02:00:00 |
+|   2500|       2405| 2016-04-27 02:00:00 |
 
 ``` r
 output$results %>% knitr::kable()
 ```
 
-| annotations.DMS.lat  | annotations.DMS.lng | annotations.MGRS | annotations.Maidenhead | annotations.Mercator.x | annotations.Mercator.y | annotations.OSM.edit\_url                                                     | annotations.OSM.url                                                                   | annotations.callingcode | annotations.geohash  | annotations.sun.rise.apparent | annotations.sun.rise.astronomical | annotations.sun.rise.civil | annotations.sun.rise.nautical | annotations.sun.set.apparent | annotations.sun.set.astronomical | annotations.sun.set.civil | annotations.sun.set.nautical | annotations.timezone.name | annotations.timezone.now\_in\_dst | annotations.timezone.offset\_sec | annotations.timezone.offset\_string | annotations.timezone.short\_name | annotations.what3words.words | bounds.northeast.lat | bounds.northeast.lng | bounds.southwest.lat | bounds.southwest.lng | components.city | components.country | components.country\_code | components.county | components.postcode | components.state | confidence | formatted                                       |  geometry.lat|  geometry.lng| components.post\_office | components.road | components.suburb | components.village |
-|:---------------------|:--------------------|:-----------------|:-----------------------|:-----------------------|:-----------------------|:------------------------------------------------------------------------------|:--------------------------------------------------------------------------------------|:------------------------|:---------------------|:------------------------------|:----------------------------------|:---------------------------|:------------------------------|:-----------------------------|:---------------------------------|:--------------------------|:-----------------------------|:--------------------------|:----------------------------------|:---------------------------------|:------------------------------------|:---------------------------------|:-----------------------------|:---------------------|:---------------------|:---------------------|:---------------------|:----------------|:-------------------|:-------------------------|:------------------|:--------------------|:-----------------|:-----------|:------------------------------------------------|-------------:|-------------:|:------------------------|:----------------|:------------------|:-------------------|
[…]
