### 2016-04-26 a6ac824
corrects time stamp issue
