### 2016-04-26 4a0d8c3
updates README and vignette to mention  Noam's improvement
