+++ README.Rmd
@@ -25,6 +25,8 @@ install_github("masalmon/opencage")
 
 To use the package, you will also need an API key. For this register at https://geocoder.opencagedata.com/pricing. The free API key provides up to 2,500 calls a day. It is recommended you save your API key as an environment variable. See https://stat545-ubc.github.io/bit003_api-key-env-var.html
 
+Both functions of the package will conveniently look for your API key using Sys.getenv("OPENCAGE_KEY") so if your API key is an environment variable called "OPENCAGE_KEY" you don't need to input it.
+
 # Geocoding
 
 The [OpenCage](https://geocoder.opencagedata.com/) API supports forward and reverse geocoding. Sources of OpenCage are open geospatial data including OpenStreetMap, Yahoo! GeoPlanet, Natural Earth Data, Thematic Mapping, Ordnance Survey OpenSpace, Statistics New Zealand, Zillow, MaxMind, GeoNames, the US Census Bureau and Flickr's shapefiles plus a whole lot more besides. See [this page](https://geocoder.opencagedata.com/credits) for the full list of credits.
@@ -37,7 +39,7 @@ Forward geocoding is from placename to latitude and longitude tuplet(s).
 
 ```{r, warning = FALSE, message = FALSE}
 library("opencage")
-output <- opencage_forward(placename = "Sarzeau", key = Sys.getenv("OPENCAGE_KEY"))
+output <- opencage_forward(placename = "Sarzeau")
 print(output$time_stamp)
 library("dplyr")
 output$rate_info %>% knitr::kable()
@@ -50,8 +52,7 @@ Reverse geocoding is from latitude and longitude to placename(s).
 
 ```{r}
 output2 <- opencage_reverse(latitude = 51.5034070, 
-                            longitude = -0.1275920, 
-                 key = Sys.getenv("OPENCAGE_KEY"))
+                            longitude = -0.1275920)
 print(output2$time_stamp)
 output2$rate_info %>% knitr::kable()
 output2$results %>% knitr::kable()
@@ -86,14 +87,11 @@ Optional parameters of both `opencage_forward` and `opencage_reverse` can make t
 Note that the package uses [memoise](https://github.com/hadley/memoise) with no timeout argument so that results are cached inside an active R session. The underlying data at OpenCage is updated about once a day.
 
 ```{r}
-system.time(opencage_reverse(latitude = 10, longitude = 10,
-key = Sys.getenv("OPENCAGE_KEY")))
+system.time(opencage_reverse(latitude = 10, longitude = 10))
 
-system.time(opencage_reverse(latitude = 10, longitude = 10,
-key = Sys.getenv("OPENCAGE_KEY")))
+system.time(opencage_reverse(latitude = 10, longitude = 10))
 
 memoise::forget(opencage_reverse)
-system.time(opencage_reverse(latitude = 10, longitude = 10,
-key = Sys.getenv("OPENCAGE_KEY")))
+system.time(opencage_reverse(latitude = 10, longitude = 10))
 
 ```
