### 2016-04-27 #36 @juliasilge
### General Comments

* :raised_hands::globe_with_meridians::raised_hands::globe_with_meridians: This is a well-executed package to call the OpenCage API for forward and reverse geocoding and, in the end, get the results in a list with a nice, convenient data frame. The functions and tests have been crafted carefully so the user interacts with the API well.
* This package has a clear, sort of minimalist scope; it has just two well-designed functions for the user to interact with and that is great.
* I have never used `memoise` before and it is spiffy! This is a great application of it.
* How to get an API key is made super clear, and the explanations of how to store and use it are also explicitly included, which is awesome. :+1:
* I am confused about how many results I will get when I query, especially for `opencage_forward()`, and what that means. When I query for Salt Lake City, I get 8 results; when I query for Fort Worth, I get 6 results. I see that the default to return is 10, but what do these multiple results mean? Are they boundary points? Are they different responses from the different sources behind OpenCage like Zillow/Yahoo/US Census/etc? (For both of those U.S. city examples, they seemed to be all from OpenStreetMap, FWIW.) I think the answer to this needs to be clear in the README, the .Rd for forward geocoding at least, and the vignette -- what do the multiple results mean? For some reason the multiple results for reverse geocoding don't seem as puzzling to me, but it still is probably good to clarify this a bit.
* The tests are thorough, clear, and well written. Almost all of the tests call the API and thus will be skipped on CRAN, but that's just the nature of the beast for this type of package, I think.

### Tests

* All tests run successfully on OSX El Capitan, R 3.2.5, after I stored an OpenCage API key as "OPENCAGE_KEY" in my .Renviron
* All code in README, vignettes, and examples knit/ran as expected (after I stored an API key)

### `DESCRIPTION` file

* The title should be title case, which means "to" and "the" should be lower case.
* The description should be a complete sentence, which only means it needs a period. Also, I believe forward and reverse should be lower case; they aren't proper nouns or anything.
* Can you make your lists for Imports and Suggests formatted the same? One package on the same line with the "Imports:|Suggests:" then one package on each following line
* This one is sooooo picky, but before using "which" like you do in the description you need a comma.

### Documentation

* The last example for `opencage_forward` is too long and is going to get truncated in the PDF manual on CRAN; you can just split that and put it on two lines like you would in an R script.
* When you are explaining how to type code in documentation (.Rd or vignette) make that more explicit. For example, for how to use `bounds` in the function documentation, say something like "of a bounding box with \\code{bounds =} min long, min lat, max long, max lat. For example, \\code{bounds = -0.563160, 51.280430, 0.278970, 51.683979}"
* Speaking of `bounds`, isn't this missing a `c()` or something? I couldn't actually get a bounding box to work. I think we should see a working example with a bounding box in the vignette.
* Use the same kind of quotes for country code and language code in your documentation, i.e. 'GB' and 'pt-BR', because these are both treated as strings. Actually, why don't you make either one or two examples in the vignette that show you using all of the parameters? It could be one API call with, like, ALL of them, just to show how they work in action.
* Either capitalize or don't capitalize all the beginning of your explanation of parameters. I actually don't know if one is preferred. May capitalizing? i.e., "Your OpenCage key", "Provides the geocoder...", "Restricts the results...", etc.
* The paragraph you are using to explain how to get an API key is super important and I am really glad you are so explicit, but it is slightly awkward with passive voice, etc. Since this is used several times, absolutely as it should be, (README, vignette, .Rds), how about some clean-up for clarity and smooth reading? Something along the lines of: "To get an API key to access OpenCage geocoding, register at https://geocoder.opencagedata.com/pricing. The free API key provides up to 2,500 calls a day. For ease of use, save your API key as an environment variable as described at https://stat545-ubc.github.io/bit003_api-key-env-var.html" Don't forget a period after this in the .Rds, but put the bit about memoise AFTER the bit about when OpenCage is updated, I think.

##### A couple of picky grammar/capitalization/punctuation things...
* In the function documentation, geocoding should not be capitalized as a proper noun, i.e., "Forward geocoding, from placename to latitude and..." I think I would remove "that is".
* In the vignette, under "Output" use a semicolon here: "details here and here; for pure conversion"
* In the vignette, at the end of "Output" either make it two sentences or use a semicolon: "one gets different columns; there can be a lot to explore!"

### Functionality

* I've been thinking about how to structure what you get from the API call, and I think that what you've done here, a list that includes a data frame, is the right choice. This makes good sense.
* Right now you have calls to `data()` in `utils.R` but a) that's not necessary because the package knows about the data files when it is loaded (I think this is how it works?!?!) and b) it's not considered good practice to do that anyway. I tried the package out after removing all of those lines and it works fine.
* You need to add the names of your data objects (`code_message`, `countrycodes`, `languagecodes`) to a call to `globalVariables` because R CMD check thinks they have "no visible bindings". I will be honest that I have no idea if this actually does anything of substance beyond quieting a `NOTE`, but if you are planning to submit to CRAN, then just go ahead. :unamused: You can see what it needs to look like [here](https://github.com/juliasilge/janeaustenr/blob/master/R/globalVariables.R); you should probably put it in a separate file like that.
* The tests in `test-opencage_query_check.R` do not `skip_on_cran()` but they call `Sys.getenv("OPENCAGE_KEY")`. How is that going to work? That needs to be changed, right?
* You need to have `#' @import memoise` in the .Rds for both main functions so that Roxygen knows that you need to import `memoise`.
