### 2016-04-28 b2207a0
adds details about multiple results
