+++ README.md
@@ -22,6 +22,8 @@ Geocoding
 
 The [OpenCage](https://geocoder.opencagedata.com/) API supports forward and reverse geocoding. Sources of OpenCage are open geospatial data including OpenStreetMap, Yahoo! GeoPlanet, Natural Earth Data, Thematic Mapping, Ordnance Survey OpenSpace, Statistics New Zealand, Zillow, MaxMind, GeoNames, the US Census Bureau and Flickr's shapefiles plus a whole lot more besides. See [this page](https://geocoder.opencagedata.com/credits) for the full list of credits.
 
+Regarding multiple results, the API doc states that "In cases where the geocoder is able to find multiple matches, the geocoder will return multiple results. The confidence or coordinates for each result should be examined to determine whether each result from an ambiguous query is sufficiently high to warrant using a result or not. A good strategy to reduce ambiguity is to use the optional `bounds` parameter described below to limit the area searched." Multiple results might mean you get a result for the airport and a road when querying a city name, or results for cities with the same name in different countries.
+
 Below are two simple examples.
 
 Forward geocoding
@@ -35,7 +37,7 @@ output <- opencage_forward(placename = "Sarzeau")
 print(output$time_stamp)
 ```
 
-    ## [1] "2016-04-26 18:30:40 CEST"
+    ## [1] "2016-04-28 11:05:10 CEST"
 
 ``` r
 library("dplyr")
@@ -44,7 +46,7 @@ output$rate_info %>% knitr::kable()
 
 |  limit|  remaining| rest                |
 |------:|----------:|:--------------------|
-|   2500|       2308| 2016-04-27 02:00:00 |
+|   2500|       2495| 2016-04-29 02:00:00 |
 
 ``` r
 output$results %>% knitr::kable()
@@ -52,8 +54,8 @@ output$results %>% knitr::kable()
 
 | annotations.DMS.lat  | annotations.DMS.lng | annotations.MGRS | annotations.Maidenhead | annotations.Mercator.x | annotations.Mercator.y | annotations.OSM.edit\_url                                                     | annotations.OSM.url                                                                   | annotations.callingcode | annotations.geohash  | annotations.sun.rise.apparent | annotations.sun.rise.astronomical | annotations.sun.rise.civil | annotations.sun.rise.nautical | annotations.sun.set.apparent | annotations.sun.set.astronomical | annotations.sun.set.civil | annotations.sun.set.nautical | annotations.timezone.name | annotations.timezone.now\_in\_dst | annotations.timezone.offset\_sec | annotations.timezone.offset\_string | annotations.timezone.short\_name | annotations.what3words.words | bounds.northeast.lat | bounds.northeast.lng | bounds.southwest.lat | bounds.southwest.lng | components.\_type | components.city | components.country | components.country\_code | components.county | components.postcode | components.state | confidence | formatted                                       |  geometry.lat|  geometry.lng| components.post\_office | components.road | components.suburb | components.village |
[…]
