+++ DESCRIPTION
@@ -1,23 +1,23 @@
 Package: opencage
 Type: Package
-Title: Interface To The OpenCage API
+Title: Interface to the OpenCage API
 Version: 0.1.0
-Authors@R: c(person("Maëlle", "Salmon", email = "maelle.salmon@yahoo.se", role = c("aut", "cre")), person("Noam", "Ross", role = "ctb"))
-Description: This packages accesses the OpenCage API which provides Forward
-    geocoding (from placename to longitude and latitude) and Reverse geocoding (from
-    longitude and latitude to placename)
+Authors@R: c(person("Maëlle", "Salmon", email = "maelle.salmon@yahoo.se", role = c("aut", "cre")), person("Noam", "Ross", role = "ctb"), person("Julia", "Silge", role = "rev"))
+Description: This packages accesses the OpenCage API, which provides forward
+    geocoding (from placename to longitude and latitude) and reverse geocoding (from
+    longitude and latitude to placename).
 License: GPL (>= 2)
 LazyData: TRUE
 URL: http://github.com/masalmon/opencage
 BugReports: http://github.com/masalmon/opencage/issues
-Imports:
-    httr,
+Imports: httr,
     jsonlite,
     dplyr,
     lubridate,
     memoise
 RoxygenNote: 5.0.1
-Suggests: testthat, lintr,
+Suggests: testthat,
+    lintr,
     knitr,
     rmarkdown
 VignetteBuilder: knitr
