### 2016-04-28 825f8f5
improves documentation
