+++ R/reverse_geocoding.R
@@ -47,24 +47,25 @@
 
 }
 
-#' Reverse Geocoding
+#' Reverse geocoding
 #'
-#' Reverse Geocoding, that is, from latitude and longitude to placename(s).
+#' Reverse geocoding, from latitude and longitude to placename(s).
 #'
-#' @param latitude latitude
-#' @param longitude longitude
-#' @param key your OpenCage key
-#' @param bounds Provides the geocoder with a hint to the region that the query resides in. This value will restrict the possible results to the supplied region. The bounds parameter should be specified as 4 coordinate points forming the south-west and north-east corners of a boundsing box. For example bounds=-0.563160,51.280430,0.278970,51.683979 (min long, min lat, max long, max lat).
+#' @param latitude Latitude
+#' @param longitude Longitude
+#' @param key Your OpenCage key
+#' @param bounds Provides the geocoder with a hint to the region that the query resides in. This value will restrict the possible results to the supplied region. The bounds parameter should be specified as 4 coordinate points forming the south-west and north-east corners of a boundsing box. For example \code{bounds = c(-0.563160, 51.280430, 0.278970, 51.683979)} (min long, min lat, max long, max lat).
 #' @param countrycode Restricts the results to the given country. The country code is a two letter code as defined by the ISO 3166-1 Alpha 2 standard. E.g. 'GB' for the United Kingdom, 'FR' for France, 'US' for United States.
-#' @param language an IETF format language code (such as es for Spanish or pt-BR for Brazilian Portuguese). If no language is explicitly specified, we will look for an HTTP Accept-Language header like those sent by a brower and use the first language specified and if none are specified en (English) will be assumed
+#' @param language An IETF format language code (such as es for Spanish or pt-BR for Brazilian Portuguese). If no language is explicitly specified, we will look for an HTTP Accept-Language header like those sent by a brower and use the first language specified and if none are specified en (English) will be assumed
 #' @param limit How many results should be returned (1-100). Default is 10.
-#' @param min_confidence an integer from 1-10. Only results with at least this confidence will be returned.
+#' @param min_confidence An integer from 1-10. Only results with at least this confidence will be returned.
 #' @param no_annotations Logical (default FALSE), when TRUE the output will not contain annotations.
 #' @param no_dedupe Logical (default FALSE), when TRUE the output will not be deduplicated
 #'
-#' @details For getting your API key register at https://geocoder.opencagedata.com/pricing. The free API key provides up to 2,500 calls a day.
-#' It is recommended you save your API key as an environment variable. See https://stat545-ubc.github.io/bit003_api-key-env-var.html
-#' Note that the package uses `memoise` with no timeout argument so that results are cached inside an active R session. The underlying data at OpenCage is updated about once a day.
+#' @details To get an API key to access OpenCage geocoding, register at https://geocoder.opencagedata.com/pricing. The free API key provides up to 2,500 calls a day. For ease of use, save your API key as an environment variable as described at https://stat545-ubc.github.io/bit003_api-key-env-var.html.
+#'  Both functions of the package will conveniently look for your API key using Sys.getenv("OPENCAGE_KEY") so if your API key is an environment variable called "OPENCAGE_KEY" you don't need to input it.
+#'
+#' The underlying data at OpenCage is updated about once a day. Note that the package uses `memoise` with no timeout argument so that results are cached inside an active R session.
 #'
[…]

[…]
