+++ R/globalVariables.R
@@ -1,0 +1,1 @@
+globalVariables(c("code_message", "countrycodes", "languagecodes"))
