### 2016-04-28 e1fd031
skips test on CRAN
