### 2016-04-28 b5afcbe
makes the key function silent by default
