### 2016-04-28 #36 @masalmon
I have a question regarding CRAN submission: the vignette will not build on CRAN, how should I make this clear?
