+++ README.Rmd
@@ -81,7 +81,7 @@ results2 <- opencage_forward(placename = "Berlin",
 results2$results %>% knitr::kable()
 ```
 
-* `countrycode`: Restricts the results to the given country. The country code is a two letter code as defined by the ISO 3166-1 Alpha 2 standard. E.g. 'GB' for the United Kingdom, 'FR' for France, 'US' for United States. See example below.
+* `countrycode`: Restricts the results to the given country. The country code is a two letter code as defined by the ISO 3166-1 Alpha 2 standard. E.g. "GB" for the United Kingdom, "FR" for France, "US" for United States. See example below.
 
 ```{r, message=FALSE}
 results3 <- opencage_forward(placename = "Berlin", country = "DE")
@@ -89,7 +89,7 @@ results3$results %>% knitr::kable()
 
 ```
 
-* `language`: an IETF format language code (such as es for Spanish or pt-BR for Brazilian Portuguese). If no language is explicitly specified, we will look for an HTTP Accept-Language header like those sent by a brower and use the first language specified and if none are specified en (English) will be assumed See example below.
+* `language`: an IETF format language code (such as "es" for Spanish or "pt-BR" for Brazilian Portuguese). If no language is explicitly specified, we will look for an HTTP Accept-Language header like those sent by a brower and use the first language specified and if none are specified en (English) will be assumed See example below.
 
 ```{r, message=FALSE}
 results3$results %>% knitr::kable()

[…]
