### 2016-04-28 #36 @sckott
no idea about `rev` - doesn't seem like a valid role, but i guess others use it?

For the vignette, don't know how to have it not build on CRAN - i cheat by building it to `.md`, then changing name to `.Rmd` - so when building, the code blocks are already rendered  - i imagine some may disagree with me though on this
