### 2016-04-28 #36 @masalmon
@sckott in which repo could I see an example of your vignette strategy/cheating? & do you comment on this in the comments when submitting?
