### 2016-04-28 #36 @masalmon
@richfitz but I guess commenting on `rev` should be enough? I should try to find another pkg where there is a `rev`.
