### 2016-04-28 #36 @richfitz
I think it's a reasonable field to have, and would be worth seeing if CRAN are willing to accept it.  They may update the fields or suggest some other way of handling this that they will tolerate.  Add it to your comments and let us know how it goes down
