### 2016-04-28 #36 @masalmon
@richfitz ok! This will be my first CRAN submission so I have to brace myself for rejection anyway.
