### 2016-04-28 #36 @noamross
The function used by `R CMD check` appears to use the whole MARC database:
https://github.com/wch/r-source/blob/e5b21d0397c607883ff25cca379687b86933d730/src/library/utils/R/citation.R#L150-L171

On Thu, Apr 28, 2016 at 9:58 AM Maëlle Salmon <notifications@github.com>
wrote:

> @richfitz <https://github.com/richfitz> ok! This will be my first CRAN
> submission so I have to brace myself for rejection anyway.
>
