### 2016-04-28 #36 @masalmon
@noamross thanks but then why do I get the NOTE on Travis and Appveyor?

```
Authors@R: c(person("Maëlle", "Salmon", email = "maelle.salmon@yahoo.se", role = c("aut", "cre")), person("Noam", "Ross", role = "ctb"), person("Julia", "Silge", role = "rev"))
```
