### 2016-04-28 #36 @richfitz
probably the `--as-cran` or other setting that make various bits come on and off on different platforms.

* the NOTE comes from here: https://github.com/wch/r-source/blob/c40eb16072ae3c8fe0f9817a20e1a37e412e9704/src/library/tools/R/QC.R#L3547-L3550
* which comes from here, if `strict` is `TRUE`: https://github.com/wch/r-source/blob/c40eb16072ae3c8fe0f9817a20e1a37e412e9704/src/library/tools/R/QC.R#L3482-L3490
* which calls this https://github.com/wch/r-source/blob/e5b21d0397c607883ff25cca379687b86933d730/src/library/utils/R/citation.R#L1321-L1335 hitting the line `role <- role[role %in% MARC_relator_db_codes_used_with_R]`
* where `MARC_relator_db_codes_used_with_R` is defined here: https://github.com/wch/r-source/blob/e5b21d0397c607883ff25cca379687b86933d730/src/library/utils/R/sysdata.R#L28-L49
