### 2016-04-28 #36 @masalmon
@richfitz well it's not hard to navigate through when someone like you gives one all useful links, many thanks! So on CRAN `strict` should be `FALSE`? That's nice.
