### 2016-04-28 #36 @richfitz
I have no idea - all I know is there are two calls to that function, one with `strict=TRUE` and one with `strict=FALSE`.  The strict one is probably triggered by running R CMD check with --as-cran, but *possibly* also involving some of the weird environment variables like `R_CHECK_CRAN_INCOMING`
