### 2016-04-29 #36 @juliasilge
This is looking SO great! :sparkles::sparkles::sparkles: The new examples in the README/vignette are really good and clear, and the package is coming together so well. I only have small changes to suggest now for consistency and readability in the documentation. Thank you so much for adding me as a reviewer in the `DESCRIPTION`; hopefully CRAN isn't annoyed on my/your account.

### README
* The paragraph about getting the API key still reads kind of awkwardly with the passive voice and all that. Would it be OK with you to use similar wording to what you currently have in the .Rds? So that whole paragraph would read like this: "This package is an interface to the OpenCage API that allows forward and reverse geocoding. To use the package, you will need an API key. To get an API key for OpenCage geocoding, register at https://geocoder.opencagedata.com/pricing. The free API key provides up to 2,500 calls a day. For ease of use, save your API key as an environment variable as described at https://stat545-ubc.github.io/bit003_api-key-env-var.html."
* Make `Sys.getenv("OPENCAGE_KEY")` code, and maybe add "manually" at the end of that sentence.
* At the beginning of the paragraph about multiple results, add one more sentence to make it clear what you're talking about: "Both forward and reverse geocoding typically return multiple results." Then you need a comma in the next sentence: "Regarding these multiple results, the API doc states, "In cases..."
* Put a comma in "For example, `bounds`"
* Don't forget the quotes for "en" in the `language` section.

### `Details` section in .Rds
* You could use `\url{}` (like `\url{https://geocoder.opencagedata.com/pricing}`) and then the URLs become clickable links.
* Don't forget a period after the sentence with the URLs, and you have two sentences right now about saving the API key and the link to Jenny Bryan's tutorial in the forward geocoding .Rd. Be sure to take out the second one.
* When you talk about `Sys.getenv("OPENCAGE_KEY")` make it `\code{}`.
* I think the discussion of multiple results is a little long here for the .Rds (but GREAT for the README and vignette). Maybe something in just one sentence like: "This function typically returns multiple results because of placename ambiguity; consider using the `bounds` parameter to limit the area searched."

### Nitpicky punctuation and spelling details for .Rds
* Periods at the end of all the descriptions of parameters in .Rds
* Comma in "For example, \\code{bounds"
* Make sure it is "bounding box", not "boundsing box"
* Don't forget the quotes for "en" in the `language` section.
* For `Value` of what is returned, either put commas at the end of all of them or don't.

### Vignette
* Take the second part of the title after the dashes and make it a subtitle.
* Do you think it would be OK to use the paragraph without passive voice, etc about getting an API key here as well? Instead of "For this register..." just start in with "To get an API key for OpenCage geocoding..."
* Make `Sys.getenv("OPENCAGE_KEY")` code.
* Make the same changes to the paragraph about multiple results as in the README.
* Put a comma in "For example, `bounds`"
* Don't forget the quotes for "en" in the `language` section.
