### 2016-04-29 #36 @juliasilge
Oh, and I have not thought or looked at examples at how to have the vignette not build on CRAN at this point.
