### 2016-04-30 #36 @jennybc
I describe how I suppress vignette building on CRAN here [in a googlesheets vignette](https://rawgit.com/jennybc/googlesheets/master/vignettes/managing-auth-tokens.html#the-making-of-googlesheets.tar.gz-for-cran).
