### 2016-04-30 #36 @masalmon
Thanks @juliasilge & @jennybc! You make improving this pkg such a pleasure! 😊
