### 2016-04-30 6c32b3d
improves README and vignette
