### 2016-04-30 8ca8402
erases double sentence
