+++ README.md
@@ -13,7 +13,7 @@ library("devtools")
 install_github("masalmon/opencage")
 ```
 
-This package is an interface to the OpenCage API that allows forward and reverse geocoding. This package is an interface to the OpenCage API that allows forward and reverse geocoding. To use the package, you will need an API key. To get an API key for OpenCage geocoding, register at <https://geocoder.opencagedata.com/pricing>. The free API key provides up to 2,500 calls a day. For ease of use, save your API key as an environment variable as described at <https://stat545-ubc.github.io/bit003_api-key-env-var.html>.
+This package is an interface to the OpenCage API that allows forward and reverse geocoding. To use the package, you will need an API key. To get an API key for OpenCage geocoding, register at <https://geocoder.opencagedata.com/pricing>. The free API key provides up to 2,500 calls a day. For ease of use, save your API key as an environment variable as described at <https://stat545-ubc.github.io/bit003_api-key-env-var.html>.
 
 Both functions of the package will conveniently look for your API key using `Sys.getenv("OPENCAGE_KEY")` so if your API key is an environment variable called "OPENCAGE\_KEY" you don't need to input it manually.
 
@@ -37,7 +37,7 @@ output <- opencage_forward(placename = "Sarzeau")
 print(output$time_stamp)
 ```
 
-    ## [1] "2016-04-30 14:50:06 CEST"
+    ## [1] "2016-04-30 14:52:11 CEST"
 
 ``` r
 library("dplyr")
@@ -46,7 +46,7 @@ output$rate_info %>% knitr::kable()
 
 |  limit|  remaining| rest                |
 |------:|----------:|:--------------------|
-|   2500|       2499| 2016-05-01 02:00:00 |
+|   2500|       2491| 2016-05-01 02:00:00 |
 
 ``` r
 output$results %>% knitr::kable()
@@ -68,7 +68,7 @@ output2 <- opencage_reverse(latitude = 51.5034070,
 print(output2$time_stamp)
 ```
 
-    ## [1] "2016-04-30 14:48:03 CEST"
+    ## [1] "2016-04-30 14:52:13 CEST"
 
 ``` r
 output2$rate_info %>% knitr::kable()
@@ -76,7 +76,7 @@ output2$rate_info %>% knitr::kable()
 
 |  limit|  remaining| rest                |
 |------:|----------:|:--------------------|
-|   2500|       2498| 2016-05-01 02:00:00 |
+|   2500|       2490| 2016-05-01 02:00:00 |
 
 ``` r
 output2$results %>% knitr::kable()
@@ -110,7 +110,7 @@ results1$results %>% knitr::kable()
 | 52° 31' 1.33140'' N  | 13° 23' 19.89564'' E | 33UUU9067619766  | JO62qm64pc             | 1490441.066            | 6860227.862            | <https://www.openstreetmap.org/edit?node=240109189#map=17/52.51704/13.38886>    | <https://www.openstreetmap.org/?mlat=52.51704&mlon=13.38886#map=17/52.51704/13.38886>   | 49                      | u33db8mmzj1t355u0c0j | 1461987360                    | 1461978000                        | 1461984960                 | 1461981900                    | 1462041180                   | 1462050660                       | 1462043520                | 1462046640                   | Europe/Berlin             | 1                                 | 7200                             | 200                                 | CEST                             | third.gurgled.such            | 52.6770365           | 13.5488599           | 52.3570365           | 13.2288599           | city              | Berlin          | Germany                  | de                       | Berlin           | 4          | Berlin, Germany                                                        |      52.51704|      13.38886| NA                | NA              |
[…]
