+++ README.md
@@ -37,7 +37,7 @@ output <- opencage_forward(placename = "Sarzeau")
 print(output$time_stamp)
 ```
 
-    ## [1] "2016-04-30 14:52:11 CEST"
+    ## [1] "2016-04-30 14:52:29 CEST"
 
 ``` r
 library("dplyr")
@@ -46,7 +46,7 @@ output$rate_info %>% knitr::kable()
 
 |  limit|  remaining| rest                |
 |------:|----------:|:--------------------|
-|   2500|       2491| 2016-05-01 02:00:00 |
+|   2500|       2439| 2016-05-01 02:00:00 |
 
 ``` r
 output$results %>% knitr::kable()
@@ -68,7 +68,7 @@ output2 <- opencage_reverse(latitude = 51.5034070,
 print(output2$time_stamp)
 ```
 
-    ## [1] "2016-04-30 14:52:13 CEST"
+    ## [1] "2016-04-30 15:00:23 CEST"
 
 ``` r
 output2$rate_info %>% knitr::kable()
@@ -76,7 +76,7 @@ output2$rate_info %>% knitr::kable()
 
 |  limit|  remaining| rest                |
 |------:|----------:|:--------------------|
-|   2500|       2490| 2016-05-01 02:00:00 |
+|   2500|       2438| 2016-05-01 02:00:00 |
 
 ``` r
 output2$results %>% knitr::kable()
@@ -110,7 +110,7 @@ results1$results %>% knitr::kable()
 | 52° 31' 1.33140'' N  | 13° 23' 19.89564'' E | 33UUU9067619766  | JO62qm64pc             | 1490441.066            | 6860227.862            | <https://www.openstreetmap.org/edit?node=240109189#map=17/52.51704/13.38886>    | <https://www.openstreetmap.org/?mlat=52.51704&mlon=13.38886#map=17/52.51704/13.38886>   | 49                      | u33db8mmzj1t355u0c0j | 1461987360                    | 1461978000                        | 1461984960                 | 1461981900                    | 1462041180                   | 1462050660                       | 1462043520                | 1462046640                   | Europe/Berlin             | 1                                 | 7200                             | 200                                 | CEST                             | third.gurgled.such            | 52.6770365           | 13.5488599           | 52.3570365           | 13.2288599           | city              | Berlin          | Germany                  | de                       | Berlin           | 4          | Berlin, Germany                                                        |      52.51704|      13.38886| NA                | NA              |
 | 44° 28' 7.21056'' N  | 71° 11' 6.27612'' W  | 19TCK2619726251  | FN44jl72sl             | -7924286.49            | 5508315.317            | <https://www.openstreetmap.org/edit?relation=170184#map=17/44.46867/-71.18508>  | <https://www.openstreetmap.org/?mlat=44.46867&mlon=-71.18508#map=17/44.46867/-71.18508> | 1                       | drvkupk8g2zdsyw72s2c | 1462008840                    | 1462001940                        | 1462006920                 | 1462004580                    | 1462060140                   | 1461980700                       | 1461975660                | 1461978060                   | America/New\_York         | 1                                 | -14400                           | -400                                | EDT                              | angle.liability.nerve         | 44.528436            | -71.1236419          | 44.445057            | -71.3978579          | city              | Berlin          | United States of America | us                       | New Hampshire    | 5          | Berlin, Coös County, New Hampshire, United States of America           |      44.46867|     -71.18508| Coös County       | NA              |
[…]
