### 2016-04-30 3b289a7
improves Rd's
