+++ vignettes/opencage.Rmd
@@ -1,5 +1,6 @@
 ---
-title: "R package for the OpenCage API -- both forward and reverse geocoding"
+title: "R package for the OpenCage API"
+subtitle: "Both forward and reverse geocoding"
 author: "Maëlle Salmon"
 date: "`r Sys.Date()`"
 output: rmarkdown::html_vignette
