### 2016-04-30 #36 @masalmon
@juliasilge I've now made all changes! Unless I forgot one comment :grimacing:. The documentation reads much better now, many thanks! You've been so patient about the bad punctuation!

Once you'll be happy with the changes I'll prepare the CRAN submission using @jennybc's and @sckott's suggestions about the vignette. :smile_cat:
