### 2016-04-30 7d3a567
Creates a companion NOT_CRAN R object in the vignette + Uses NOT_CRAN via chunk options to conditionally suppress extraction and compilation of vignette chunks that are not CRAN safe.
