### 2016-05-03 #36 @masalmon
Thank you so much, @juliasilge! I will make this correction.

@sckott should my next step be to transfer the repo to ropenscilabs or to submit to CRAN? I won't do it before the week-end anyway I think.

An issue I had was my first name. I wonder which encoding is better for the description file to still be "Maëlle" so I want to check various solutions. When I tried a few days ago to build source and then look at the DESCRIPTION in the .tar.gz my name was Ma?lle or other variations. That said I'm Maëlle [in this package](https://cran.r-project.org/web/packages/surveillance/index.html) so I'll look at this description file. [I sound so narcissistic now...]
