### 2016-05-03 #36 @sckott
> An issue I had was my first name.

I think you want `Encoding: UTF-8` in your DESCRIPTION file
