### 2016-05-03 #36 @masalmon
oooh many thanks!!

    Den tisdag, 3 maj 2016 18:50 skrev Scott Chamberlain <notifications@github.com>:

An issue I had was my first name.
I think you want Encoding: UTF-8 in your DESCRIPTION file—
You are receiving this because you authored the thread.
Reply to this email directly or view it on GitHub
