### 2016-05-03 #36 @masalmon
Yes it did! This line + making sure I saved it with UTF-8 encoding in RStudio. Thank you!
