### 2016-05-03 574db33
changes the DESCRIPTION encoding and ignores inst/
