### 2016-05-03 b3496e4
Makes the title and subtitle in the vignette title case
