### 2016-05-03 #36 @sckott
@masalmon

> should my next step be to transfer the repo to ropenscilabs or to submit to CRAN?

Transfer to `ropenscilabs` - then submit to CRAN whenever you're ready - (a pkg can be on CRAN already when submitted here of course)
