+++ README.md
@@ -10,7 +10,7 @@ To install the package, you will need the devtools package.
 
 ``` r
 library("devtools")
-install_github("masalmon/opencage")
+install_github("ropenscilabs/opencage")
 ```
 
 This package is an interface to the OpenCage API that allows forward and reverse geocoding. To use the package, you will need an API key. To get an API key for OpenCage geocoding, register at <https://geocoder.opencagedata.com/pricing>. The free API key provides up to 2,500 calls a day. For ease of use, save your API key as an environment variable as described at <https://stat545-ubc.github.io/bit003_api-key-env-var.html>.
@@ -37,7 +37,7 @@ output <- opencage_forward(placename = "Sarzeau")
 print(output$time_stamp)
 ```
 
-    ## [1] "2016-04-30 14:52:29 CEST"
+    ## [1] "2016-05-03 20:40:42 CEST"
 
 ``` r
 library("dplyr")
@@ -46,7 +46,7 @@ output$rate_info %>% knitr::kable()
 
 |  limit|  remaining| rest                |
 |------:|----------:|:--------------------|
-|   2500|       2439| 2016-05-01 02:00:00 |
+|   2500|       2380| 2016-05-04 02:00:00 |
 
 ``` r
 output$results %>% knitr::kable()
@@ -54,8 +54,8 @@ output$results %>% knitr::kable()
 
 | annotations.DMS.lat  | annotations.DMS.lng | annotations.MGRS | annotations.Maidenhead | annotations.Mercator.x | annotations.Mercator.y | annotations.OSM.edit\_url                                                     | annotations.OSM.url                                                                   | annotations.callingcode | annotations.geohash  | annotations.sun.rise.apparent | annotations.sun.rise.astronomical | annotations.sun.rise.civil | annotations.sun.rise.nautical | annotations.sun.set.apparent | annotations.sun.set.astronomical | annotations.sun.set.civil | annotations.sun.set.nautical | annotations.timezone.name | annotations.timezone.now\_in\_dst | annotations.timezone.offset\_sec | annotations.timezone.offset\_string | annotations.timezone.short\_name | annotations.what3words.words | bounds.northeast.lat | bounds.northeast.lng | bounds.southwest.lat | bounds.southwest.lng | components.\_type | components.city | components.country | components.country\_code | components.county | components.postcode | components.state | confidence | formatted                                       |  geometry.lat|  geometry.lng| components.post\_office | components.road | components.suburb | components.village |
 |:---------------------|:--------------------|:-----------------|:-----------------------|:-----------------------|:-----------------------|:------------------------------------------------------------------------------|:--------------------------------------------------------------------------------------|:------------------------|:---------------------|:------------------------------|:----------------------------------|:---------------------------|:------------------------------|:-----------------------------|:---------------------------------|:--------------------------|:-----------------------------|:--------------------------|:----------------------------------|:---------------------------------|:------------------------------------|:---------------------------------|:-----------------------------|:---------------------|:---------------------|:---------------------|:---------------------|:------------------|:----------------|:-------------------|:-------------------------|:------------------|:--------------------|:-----------------|:-----------|:------------------------------------------------|-------------:|-------------:|:------------------------|:----------------|:------------------|:-------------------|
[…]
