+++ README.Rmd
@@ -19,7 +19,7 @@ To install the package, you will need the devtools package.
 
 ```{r, eval = FALSE}
 library("devtools")
-install_github("masalmon/opencage")
+install_github("ropenscilabs/opencage")
 
 ```
 
@@ -120,3 +120,13 @@ memoise::forget(opencage_reverse)
 system.time(opencage_reverse(latitude = 10, longitude = 10))
 
 ```
+
+
+## Meta
+
+* Please [report any issues or bugs](https://github.com/ropenscilabs/opencage/issues).
+* License: GPL
+* Get citation information for `opencage` in R doing `citation(package = 'opencage')`
+* Please note that this project is released with a [Contributor Code of Conduct](CONDUCT.md). By participating in this project you agree to abide by its terms.
+
+[![ropensci_footer](http://ropensci.org/public_images/github_footer.png)](http://ropensci.org)
