### 2016-05-03 #36 @sckott
great, thanks!

> I'd like to let this issue open as long as I haven't got the package on CRAN.

Will that be soon?  We generally close these issues after approval and repo transfer, but if it's soon we can leave this open until then
